+++ app/components/FireBaseMessaging.js
@@ -1,17 +1,18 @@
 import React, { Component, Fragment } from 'react';
-import { Alert } from 'react-native';
+import { Alert, Platform, AppState, AppStateStatus, Linking } from 'react-native';
 import PropTypes from 'prop-types';
 import * as firebase from 'react-native-firebase';
 import { connect } from 'react-redux';
 import _ from 'lodash';
 import PushNotificationIOS from '@react-native-community/push-notification-ios';
-import { Platform, AppState, AppStateStatus } from 'react-native';
 
+import { Actions } from 'react-native-router-flux';
+import moment from 'moment';
 import { setFcmToken } from '../state/fcm/fcm.actions';
 import { appletsSelector } from '../state/applets/applets.selectors';
 import { setCurrentApplet } from '../state/app/app.actions';
 import { startResponse } from '../state/responses/responses.thunks';
-import { sync } from '../state/app/app.thunks';
+import { inProgressSelector } from '../state/responses/responses.selectors';
 
 const AndroidChannelId = 'MindLoggerChannelId';
 const fMessaging = firebase.messaging.nativeModuleExists && firebase.messaging();
@@ -34,13 +35,16 @@ class FireBaseMessaging extends Component {
     this.notificationDisplayedListener = fNotifications
       .onNotificationDisplayed(this.onNotificationDisplayed);
     this.notificationListener = fNotifications.onNotification(this.onNotification);
-    this.notificationOpenedListener = fNotifications.onNotificationOpened(this.onNotificationOpened);
+    this.notificationOpenedListener = fNotifications
+      .onNotificationOpened(this.onNotificationOpened);
     this.onTokenRefreshListener = fMessaging.onTokenRefresh(this.onTokenRefresh);
     this.messageListener = fMessaging.onMessage(this.onMessage);
 
     fMessaging.hasPermission().then((granted) => {
       if (!granted) {
-        fMessaging.requestPermission();
+        fMessaging.requestPermission()
+          .then(() => this.checkPermissionAgain())
+          .catch(() => this.checkPermissionAgain());
       }
     });
 
@@ -53,62 +57,120 @@ class FireBaseMessaging extends Component {
 
     if (isIOS) {
       await firebase.messaging().ios.registerForRemoteNotifications();
-      const apns = await firebase.messaging().ios.getAPNSToken();
+      // const apns = await firebase.messaging().ios.getAPNSToken();
 
       // eslint-disable-next-line no-console
-      console.log(`FCM[${Platform.OS}] APNSToken: ${apns}`);
+      // console.log(`FCM[${Platform.OS}] APNSToken: ${apns}`);
     }
   }
 
   componentWillUnmount() {
-    this.notificationDisplayedListener();
-    this.notificationListener();
-    this.notificationOpenedListener();
-    this.onTokenRefreshListener();
-    this.messageListener();
-
-    AppState.addEventListener('change', this.handleAppStateChange);
-  }
-
-  openActivityByEventId = notificationObj => {
-    const eventId = _.get(notificationObj, 'notification._data.event_id', '');
-    if (eventId) {
-      this.props.sync(() => this.openActivityByEventIdCb(eventId));
+    if (this.notificationDisplayedListener) {
+      this.notificationDisplayedListener();
     }
+    if (this.notificationListener) {
+      this.notificationListener();
+    }
+    if (this.notificationOpenedListener) {
+      this.notificationOpenedListener();
+    }
+    if (this.onTokenRefreshListener) {
+      this.onTokenRefreshListener();
+    }
+    if (this.messageListener) {
+      this.messageListener();
+    }
+
+    AppState.removeEventListener('change', this.handleAppStateChange);
   }
 
-  openActivityByEventIdCb = eventId => {
-    let schema = null;
-    const currentApplet = this.props.applets.find(({ schedule: { events } }) => {
-      const event = events.find(({ id }) => id === eventId);
-      if (event) {
-        schema = event.data.URI;
+  checkPermissionAgain = () => {
+    fMessaging.hasPermission().then((granted) => {
+      if (!granted) {
+        Alert.alert(
+          '"MindLogger" would like to send you notifications',
+          'These can be configured in Settings',
+          [
+            {
+              text: 'Dismiss',
+              style: 'cancel',
+            },
+            {
+              text: 'Open Settings',
+              onPress: this.openSettings,
+              style: 'default',
+            },
+          ],
+        );
       }
-      return event;
     });
-    if (!currentApplet) {
-      Alert.alert('Applet was not found', 'There is no applet for given event id.');
-      return;
-    }
+  }
 
-    const currentActivity = currentApplet.activities.find(activity => activity.schema === schema);
-    if (!currentActivity) {
-      Alert.alert('Activity was not found', 'There is no activity for given event id.');
-      return;
-    }
+  openSettings = () => Linking.openSettings();
 
-    this.props.setCurrentApplet(currentApplet.id);
-    this.props.startResponse(currentActivity);
+  isCompleted = activity => activity.lastResponseTimestamp !== null
+    && activity.nextScheduledTimestamp === null
+    && (moment().isSame(moment(activity.lastResponseTimestamp), 'day'));
+
+  isActivityCompleted = (currentApplet, currentActivity) => {
+    const inProgressKeys = Object.keys(this.props.inProgress);
+    if (inProgressKeys) {
+      const isActivityNotInProgress = !inProgressKeys
+        .includes(currentApplet.id + currentActivity.id);
+      if (isActivityNotInProgress) {
+        return this.isCompleted(currentActivity);
+      }
+      return false;
+    }
+    return this.isCompleted(currentActivity);
+  };
+
+  openActivityByEventId = (notificationObj) => {
+    const eventId = _.get(notificationObj, 'notification._data.event_id', '');
+    const appletId = _.get(notificationObj, 'notification._data.applet_id', '');
+    const activityId = _.get(notificationObj, 'notification._data.activity_id', '');
+    // eslint-disable-next-line no-console
+    console.log('openActivityByEventId', { eventId, appletId, activityId });
+
+    if (eventId && appletId && activityId) {
+      const currentApplet = this.props.applets.find(applet => applet.id === `applet/${appletId}`);
+      if (!currentApplet) {
+        Alert.alert('Applet was not found', 'There is no applet for given id.');
+        return;
+      }
+      const currentActivity = currentApplet.activities.find(activity => activity.id === `activity/${activityId}`);
+      if (!currentActivity) {
+        Alert.alert('Activity was not found', 'There is no activity for given id.');
+        return;
+      }
+
+      const isActivityCompleted = this.isActivityCompleted(currentApplet, currentActivity);
+      this.props.setCurrentApplet(`applet/${appletId}`);
+
+      if (isActivityCompleted) {
+        Actions.push('applet_details', { initialTab: 'data' });
+        Alert.alert('', `You have already completed ‘${currentActivity.name.en}’`);
+        return;
+      }
+      Actions.push('applet_details');
+
+      if (new Date().getTime() - (currentActivity.nextScheduledTimestamp?.getTime() ?? 0) >= 0) {
+        this.props.startResponse(currentActivity);
+      } else {
+        const time = moment(currentActivity.nextScheduledTimestamp).format('HH:mm');
+        Alert.alert('Activity not ready', `You’re not able to start activity yet, ‘${currentActivity.name.en}’ is scheduled to start at ${time} today`);
+      }
+    }
   }
 
   handleAppStateChange = (nextAppState: AppStateStatus) => {
     const isAppStateChanged = this.state.appState !== nextAppState;
-
     if (isAppStateChanged) {
-      firebase.notifications().setBadge(0);
+      this.setState({ appState: nextAppState });
+      if (isIOS) {
+        this.updateApplicationIconBadgeNumber();
+      }
     }
-
-    this.setState({ appState: nextAppState });
   }
 
   initAndroidChannel = () => {
@@ -127,12 +189,29 @@ class FireBaseMessaging extends Component {
     }
   };
 
-  onNotificationDisplayed = (notification: firebase.RNFirebase.notifications.Notification) => {
+  onNotificationDisplayed = async (
+    notification: firebase.RNFirebase.notifications.Notification,
+  ) => {
     // eslint-disable-next-line no-console
     console.log(`FCM[${Platform.OS}]: onNotificationDisplayed`, notification);
+    if (isIOS) {
+      await this.updateApplicationIconBadgeNumber();
+    }
   };
 
-  onNotification = (notification: firebase.RNFirebase.notifications.Notification) => {
+  getDeliveredNotificationsCount = ():Promise<number> => new Promise((resolve) => {
+    PushNotificationIOS.getDeliveredNotifications((notifications) => {
+      resolve(notifications.length);
+    });
+  });
+
+  getScheduledLocalNotificationsCount = ():Promise<number> => new Promise((resolve) => {
+    PushNotificationIOS.getScheduledLocalNotifications((notifications) => {
+      resolve(notifications.length);
+    });
+  });
+
+  onNotification = async (notification: firebase.RNFirebase.notifications.Notification) => {
     // eslint-disable-next-line no-console
     console.log('onNotification', { notification });
 
@@ -142,8 +221,13 @@ class FireBaseMessaging extends Component {
       subtitle: notification.subtitle,
       body: notification.body,
       data: notification.data,
-      iosBadge: notification.ios.badge,
+      // iosBadge: notification.ios.badge,
     });
+
+    if (isIOS) {
+      const notificationsCount = await this.generateApplicationIconBadgeNumber();
+      localNotification.ios.setBadge(notificationsCount);
+    }
 
     firebase.notifications().displayNotification(localNotification).catch((error) => {
       // eslint-disable-next-line no-console
@@ -151,12 +235,28 @@ class FireBaseMessaging extends Component {
     });
   };
 
-  onNotificationOpened = (notificationOpen: firebase.RNFirebase.notifications.NotificationOpen) => {
+  onNotificationOpened = async (
+    notificationOpen: firebase.RNFirebase.notifications.NotificationOpen,
+  ) => {
     // eslint-disable-next-line no-console
     this.openActivityByEventId(notificationOpen);
 
+    // eslint-disable-next-line no-console
     console.log(`FCM[${Platform.OS}]: onNotificationOpened `, notificationOpen);
-    firebase.notifications().setBadge(0);
+    if (isIOS) {
+      await this.updateApplicationIconBadgeNumber();
+    }
+  };
+
+  updateApplicationIconBadgeNumber = async () => {
+    const iconBadgeNumber = await this.generateApplicationIconBadgeNumber();
+    PushNotificationIOS.setApplicationIconBadgeNumber(iconBadgeNumber);
+  }
+
+  generateApplicationIconBadgeNumber = async () => {
+    const deliveredNotificationsCount = await this.getDeliveredNotificationsCount();
+    const scheduledLocalNotificationsCount = await this.getScheduledLocalNotificationsCount();
+    return deliveredNotificationsCount + scheduledLocalNotificationsCount;
   };
 
   onTokenRefresh = (fcmToken: string) => {
@@ -166,27 +266,27 @@ class FireBaseMessaging extends Component {
     setFCMToken(fcmToken);
   };
 
-  onMessage = (message: firebase.RNFirebase.messaging.RemoteMessage) => {
+  onMessage = async (message: firebase.RNFirebase.messaging.RemoteMessage) => {
     // eslint-disable-next-line no-console
     console.log(`FCM[${Platform.OS}]: onMessage: `, message, message.data);
     // eslint-disable-next-line no-console
     console.log(`FCM[${Platform.OS}]: message.data: ${message.data}`);
     const { data } = message;
 
-    PushNotificationIOS.getApplicationIconBadgeNumber((prevBadges = 0) => {
-      const localNotification = this.newNotification({
-        notificationId: message.messageId,
-        title: data.title || 'Push Notification',
-        subtitle: data.subtitle || null,
-        data,
-        iosBadge: prevBadges + 1,
-      });
-
-      firebase.notifications().displayNotification(localNotification).catch((error) => {
-        // eslint-disable-next-line no-console
-        console.warn(`FCM[${Platform.OS}]: error `, error);
-      });
-
+    const localNotification = this.newNotification({
+      notificationId: message.messageId,
+      title: data.title || 'Push Notification',
+      subtitle: data.subtitle || null,
+      data,
+      // iosBadge: prevBadges + 1,
+    });
+    if (isIOS) {
+      const notificationsCount = await this.generateApplicationIconBadgeNumber();
+      localNotification.ios.setBadge(notificationsCount + 1);
+    }
+    firebase.notifications().displayNotification(localNotification).catch((error) => {
+      // eslint-disable-next-line no-console
+      console.warn(`FCM[${Platform.OS}]: error `, error);
     });
   };
 
@@ -226,10 +326,15 @@ class FireBaseMessaging extends Component {
 FireBaseMessaging.propTypes = {
   children: PropTypes.node.isRequired,
   setFCMToken: PropTypes.func.isRequired,
+  applets: PropTypes.array.isRequired,
+  inProgress: PropTypes.array.isRequired,
+  setCurrentApplet: PropTypes.func.isRequired,
+  startResponse: PropTypes.func.isRequired,
 };
 
 const mapStateToProps = state => ({
-  applets: appletsSelector(state)
+  applets: appletsSelector(state),
+  inProgress: inProgressSelector(state),
 });
 
 const mapDispatchToProps = dispatch => ({
@@ -238,7 +343,6 @@ const mapDispatchToProps = dispatch => ({
   },
   setCurrentApplet: id => dispatch(setCurrentApplet(id)),
   startResponse: activity => dispatch(startResponse(activity)),
-  sync: cb => dispatch(sync(cb))
 });
 
 export default connect(mapStateToProps, mapDispatchToProps)(FireBaseMessaging);