+++ app/components/core/TouchBox.js
@@ -2,7 +2,10 @@ import React, { useState, useEffect } from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import { StyleSheet, View, TouchableOpacity } from 'react-native';
-import { getResponseInActivity, getResponseInApplet } from '../../state/responses/responses.actions';
+import {
+  getResponseInActivity,
+  getResponseInApplet,
+} from '../../state/responses/responses.actions';
 import { colors } from '../../theme';
 
 const styles = StyleSheet.create({
@@ -43,7 +46,15 @@ const styles = StyleSheet.create({
 
 // const TouchableOpacityEx = withPreventDoubleClick(TouchableOpacity);
 
-const TouchBox = ({ children, activity, onPress, getResponseInActivity, getResponseInApplet, isActivity, isApplet }) => {
+const TouchBox = ({
+  children,
+  activity,
+  onPress,
+  getResponseInActivity,
+  getResponseInApplet,
+  isActivity,
+  isApplet,
+}) => {
   const [touched, setTouched] = useState(false);
   const [appletTouched, setAppletTouched] = useState(false);
 
@@ -79,10 +90,13 @@ const TouchBox = ({ children, activity, onPress, getResponseInActivity, getRespo
   }, [isApplet]);
 
   return (
-    <TouchableOpacity disabled={(activity && activity.status === 'scheduled' && !activity.nextAccess)} onPress={handlePress}>
-      <View style={styles.box}>
-        {children}
-      </View>
+    <TouchableOpacity
+      disabled={
+        activity && activity.status === 'scheduled' && !activity.nextAccess
+      }
+      onPress={handlePress}
+    >
+      <View style={styles.box}>{children}</View>
     </TouchableOpacity>
   );
 };
@@ -107,4 +121,7 @@ const mapDispatchToProps = {
   getResponseInApplet,
 };
 
-export default connect(mapStateToProps, mapDispatchToProps)(TouchBox);
+export default connect(
+  mapStateToProps,
+  mapDispatchToProps,
+)(TouchBox);